fix: enhance resource group name validation logic for spaces and special characters (#2916)
### This PR resolves [#2915](https://github.com/lablup/backend.ai-webui/issues/2915) issue

Related PR: [control-panel#930](https://github.com/lablup/backend.ai-control-panel/pull/930)

**Changes:**

Currently, validation rules are not applied in Backend.AI Core, but plan to add them to the scope of domain and project in the future. Therefore, I added validation rules (allow dot and unicode) to match the rules on the core side.

**Invalid strings:**
- contains special characters
- consisting of leading, trailing, or only spaces, or more than one space between characters.
- empty string

**How to test:**
- Move to resources page and click resource group tab
- Create a new resource group and verify that the validation rule works correctly.