feat(FR-691): Add description experimental setting group (#3391)
Resolves #3392 (FR-691)

### Add description to Experimental Features section in User Settings

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XqC2uNFuj0wg8I60sMUh/d4575947-774c-4df8-9a9a-aeb468400821.png)

This PR adds a descriptive text to the Experimental Features section in the User Settings page to better inform users about the nature of these features. The description explains that these are new features available for testing before official release and that they may change or be removed in future updates.

The changes include:
- Adding a new `description` field to the settings group type in UserSettingsPage.tsx
- Adding the description text for the Experimental Features section
- Adding translations for the new description text in all supported languages