fix(FR-242): show vfolder list based on delegated user (#3070)
resolves #3017 (FR-242)
<!-- replace NNN, MMM with the GitHub issue number and the corresponding Jira issue number. -->

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

> The ability to create a session by setting a Session Owner is not currently available. https://lablup.atlassian.net/browse/BA-414
> In this PR, I'm aiming to show the right vfolder list for the set session owner.

In the `VFolderTable` used by the session launcher, we are fetching the vfolder list via the REST API. Since the API provides the ability to filter the data based on owner email, use it to show the vfolder list based on the delegated owner email assigned by the session launcher.

**feature:**
- If owner is set in the session launcher, it will show a list of vfolders for that user.
- If a previously selected vfolder exists, all selections will be reset when the owner related settings are changed.

**Checklist:** (if applicable)

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after