fix(FR-823): log message scroll bar is too wide when logs are empty. (#3487)
Resolves #3486 ([FR-823](https://lablup.atlassian.net/browse/FR-823))

# Conditionally set table scroll height in ErrorLogList

This PR modifies the `ErrorLogList` component to conditionally set the vertical scroll height. When there are no error logs to display (filtered log data with `isError` is empty), the vertical scroll height is set to `undefined` instead of a fixed calculation. This prevents unnecessary scrollbars from appearing when there's no error data to scroll through.

**Checklist:**

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after

[FR-823]: https://lablup.atlassian.net/browse/FR-823?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ