fix(FR-921): Fix incorrect access key when terminating another user's session (#3619)
resolves #3596 (FR-921)

This PR updates the `TerminateSessionModal` component to use the session's access key directly from the session object instead of using the global access key from the BAI client configuration. This change ensures that the correct access key is used when terminating a session, particularly in cases where the session might have been created with a different access key than the one currently configured.

**Checklist:**

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after