refactor(FR-844): migrate Terms of Service modal to react component (#3656)
resolves #3512 [(FR-844)](https://lablup.atlassian.net/browse/FR-844)
<!-- replace NNN, MMM with the GitHub issue number and the corresponding Jira issue number. -->

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->
Migrate the Terms of Service to a React modal in that PR. This modal is only active in parts that are written in React.

**changes**
* add `TermsOfServiceModal.tsx`
* The modal has a width of `80%` and I'm not sure if that value is appropriate.

You can enable the modal by clicking on Terms of Service at the bottom of the side menu.

![CleanShot 2025-05-14 at 11.51.24@2x.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xbCemO1RqqcSjEXCRK3p/642284a6-05b9-490a-bb06-7c43e9b6309e.png)

**Checklist:** (if applicable)

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after