refactor(FR-992): migrate Privacy Policy modal to NEO Style react component (#3658)
resolves #NNN (FR-992)
<!-- replace NNN, MMM with the GitHub issue number and the corresponding Jira issue number. -->

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

Migrate the `Privacy Policy` modal from that PR to a React component in NEO Style. This modal is only visible in React.

**changes**
* add `PrivacyPolicyModal`
* If the language is anything other than Korean, it will be displayed in English.

You can view this modal by clicking the `privacy policy` button at the bottom of the side menu.

![CleanShot 2025-05-14 at 13.16.20@2x.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xbCemO1RqqcSjEXCRK3p/c9ec3597-1676-47f9-80ff-99403205550d.png)

**Checklist:** (if applicable)

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after