fix(FR-994): Validate that TCP APPs are allowed when open app laucnher on session detail panel (#3663)
resolves #3660 (FR-994)
<!-- replace NNN, MMM with the GitHub issue number and the corresponding Jira issue number. -->

### Fixed an issue where port information was not displayed correctly in connected info when running an ssh/sftp session from the session detail panel.

The app launcher running in the session detail panel is using the `backend-ai-app-launcher`. The `backend-ai-app-launcher` validates the tcp-based app spawn eligibility when the app launcher dialog is displayed on the screen. However, this process is missing when opening the app launcher in the session detail panel.

**Bug report:**
When refreshing after creating a new session but before the APP DIALOG was displayed, the port information was not displayed in the INFORMATION MODAL when executing SSH/SFTP through the SESSION DETAIL PANEL.

**How to test:**
- Connect to a test server with SSH/SFTP enabled (e.g. dogbowl).
- Create a new session and refresh the browser before opening the app launcher dialog
- Open the `session detail panel` and run ssh/sftp
- Verify that information modal shows correct `user/host/port`

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after