fix(FR-1056): Use fetch-event-source for authenticated SSE requests (#3774)
Resolves #3740 ([FR-1056](https://lablup.atlassian.net/browse/FR-1056))

<!-- replace NNN, MMM with the GitHub issue number and the corresponding Jira issue number. -->

Currently, when accessing an endpoint on a remote server, the authentication cookie is not included in the request, making background task tracking impossible due to permissions issues.

By default, when you declare an SSE handler, you can declare a BaseType that will be common to all handlers.
```
const SSEEventHandlers: SSEEventHandlerTypes<BackgroundTaskEvent> = { ... }
```
If you need a different type for a particular handler, you can override it by declaring the handler that requires type overriding next to the BaseType.
```
const SSEEventHandlers: SSEEventHandlerTypes<BackgroundTaskEvent, { onFailed: {message: {error: string}}}>
```

This PR fixes this by sending requests with the session ID in the request header so that authentication to the background task can be tracked properly.

**Changes:**
- Instead of using the EventSource web api (EventSource does not support the ability to set custom headers), modify the request to include the authentication session ID in a custom header using microsoft's fetch-event-source library.

**How to test:**
- Set the remote server to endpoint and log in to webui.
- Perform a task that has a background task (e.g. rescan images, validate services, etc.)
- Verify that SSE is coming through.

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after

[FR-1056]: https://lablup.atlassian.net/browse/FR-1056?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ