chore(FR-1197): modify validation message in environments select (#3897)
resolves #3896 (FR-1197)

## description
In the "Environments & Resource allocation" step of the session page, the validation message for the Environments select item has been updated. Previously, the validation relied on the default behavior provided by antd, but now a custom message has been specified in the rules to display a more appropriate text. The updated messages are confi
rmed to appear correctly in both Korean, English and other languages.

<img width="450" alt="Screenshot 2025-07-04 at 15 40 34" src="https://github.com/user-attachments/assets/d4831a1a-499d-4931-b44b-ce40855dab8e" />

<img width="450" alt="Screenshot 2025-07-04 at 15 41 07" src="https://github.com/user-attachments/assets/070d2a9f-eec4-438e-affb-184491d01b70" />

<br><br><br>

**Checklist:** (if applicable)

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after