feat(FR-1107): Apply BAIBoard to customize prometheus metrics display (#3856)
resolves #3810 (FR-1107)
<!-- replace NNN, MMM with the GitHub issue number and the corresponding Jira issue number. -->

> **!Notice**
There is a bug with changing the vertical size of a card. If you increase the card vertically and then decrease the size vertically again, the graphs inside the card do not shrink.

This PR implements a draggable and resizable metrics board in the User Sessions Metrics page. Users can now customize the layout of their metrics graphs by dragging and resizing them according to their preferences. The layout is persisted in local storage so it's maintained between sessions.

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/lSyr8xXz1wdXALkJKzVx/4cc01643-5af0-4d07-b498-b84cf4fd43bb.png)

<!--
Please precisely, concisely, and concretely describe what this PR changes, the rationale behind codes,
and how it affects the users and other developers.
-->

**Checklist:** (if applicable)

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after