feat(FR-873): set dashboard page as a default (#3554)
Resolves #3546 ([FR-873](https://lablup.atlassian.net/browse/FR-873))

## Rename `experimental_dashboard` to `classic_dashboard_page` with inverted logic

This PR renames the user setting `experimental_dashboard` to `classic_dashboard_page` and inverts its logic. The dashboard page is now the default view, and users can opt to use the classic summary page instead.

The change affects:
- Router configuration in App.tsx
- WebUISider menu item display logic
- User settings definition and UI controls
- All language translation files with updated descriptions

This change better reflects that the dashboard is now the standard view rather than an experimental feature, while still allowing users to switch back to the classic summary page if preferred.

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/2HueYSdFvL8pOB5mgrUQ/83d03b46-701d-484f-97e4-73da95fe446c.png)

**Checklist:**
- [x] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after

[FR-873]: https://lablup.atlassian.net/browse/FR-873?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ