fix(FR-1644): modify to prevent deletion of invited folders (#4579)
resolves #4577(FR-1644)

This PR adds a new `BAIVFolderDeleteButton` component that checks folder permissions before allowing deletion. The component:

- Checks if the user has `delete_vfolder` permission for the selected folders
- Applies appropriate styling based on permission status
- Shows a warning when some folders can't be deleted due to permission issues

The PR also updates the folder deletion modal to filter out folders that the user doesn't have permission to delete, and adds a notification to inform users when some folders are excluded from deletion.

**Checklist:** (if applicable)

- [x] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after